+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.0-dev.3
+- make it possible to omit Request/Response types 
+    (**only String is currently supported as an alternative return type**)
+- update package description in pubspec.yaml
+
 ## 0.0.0-dev.2
 - update pubspec repo url